+++ CHANGELOG.md
@@ -4,6 +4,7 @@
  * Storm UI now uses nimbus.host to find Nimbus rather than always using localhost (thanks Frostman)
  * Added report-error! to Clojure DSL
  * Automatically throttle errors sent to Zookeeper/Storm UI when too many are reported in a time interval (all errors are still logged) Configured with TOPOLOGY_MAX_ERROR_REPORT_PER_INTERVAL and TOPOLOGY_ERROR_THROTTLE_INTERVAL_SECS
+ * Kryo instance used for serialization can now be controlled via IKryoFactory interface and TOPOLOGY_KRYO_FACTORY config
  
 ## 0.8.1